Fixed minor bugs (such as ball randomly going NaN) and adding redundancy to ensure game is very smooth. Split applyBonusEffect into different functions, added fun colored text, sound effects, polish,and updated readme file.

+++ src/main/java/brickGame/GameEngine.java
@@ -20,7 +20,7 @@ public class GameEngine {
 
     public void setFps(int fps) {
         if (fps > 0) {
-            this.frameTimeMillis = 1000 / fps;
+            this.frameTimeMillis = (int) (1000.0 / fps);
         } else {
             throw new IllegalArgumentException("FPS must be greater than 0");
         }
